Updated AddBottomSheetContent -> EditItemBottomSheetContent to used BaseItem rather than DefaultItem, so that it can be used on ListScreen as well to edit Items added to ItemList. Separated Item data classes into their own files. Quantity now updates to 1.0 if left at 0.0.
