+++ app/src/main/java/com/heyzeusv/yourlists/database/models/Item.kt
@@ -2,18 +2,8 @@ package com.heyzeusv.yourlists.database.models
 
 import androidx.room.Entity
 import androidx.room.ForeignKey
-import androidx.room.Fts4
 import androidx.room.Index
 import androidx.room.PrimaryKey
-
-interface BaseItem {
-    val itemId: Long
-    val name: String
-    val category: String
-    val quantity: Double
-    val unit: String
-    val memo: String
-}
 
 @Entity(
     foreignKeys = [
@@ -53,49 +43,27 @@ data class Item(
     override val unit: String,
     override val memo: String,
     val parentItemListId: Long,
-    val originItemListId: Long? = null
-) : BaseItem
-
-@Entity(
-    foreignKeys = [ForeignKey(
-        entity = Category::class,
-        parentColumns = ["name"],
-        childColumns = ["category"],
-        onDelete = ForeignKey.RESTRICT,
-        onUpdate = ForeignKey.CASCADE
-    )],
-    indices = [Index(
-        value = ["category"],
-        name = "index_default_category_name"
-    )]
-)
-data class DefaultItem(
-    @PrimaryKey(autoGenerate = true)
-    override val itemId: Long = 0L,
-    override val name: String = "",
-    override val category: String = "",
-    override val quantity: Double = 0.0,
-    override val unit: String = "",
-    override val memo: String = "",
+    val originItemListId: Long? = null,
 ) : BaseItem {
 
-    fun toItem(parentItemListId: Long): Item {
+    override fun editCopy(
+        itemId: Long,
+        name: String,
+        category: String,
+        quantity: Double,
+        unit: String,
+        memo: String
+    ): Item {
         return Item(
             itemId = itemId,
             name = name,
-            isChecked = false,
+            isChecked = this.isChecked,
             category = category,
             quantity = quantity,
             unit = unit,
             memo = memo,
-            parentItemListId = parentItemListId
+            parentItemListId = this.parentItemListId,
+            originItemListId = this.originItemListId,
         )
     }
 }
-
-@Fts4(contentEntity = DefaultItem::class)
-@Entity
-data class DefaultItemFts(
-    val name: String,
-    val category: String
-)
